+++ Changelog.md
@@ -4,6 +4,8 @@
 
 * CABAL_DISABLE_DEPS env var to disable dependencies install by cabal. This can
   be useful when we have dependencies already installed e.g. in a nix shell.
+* Add support for working with remote repositories.
+  - Add `GIT_SOURCE`, `GIT_REV`, and `GIT_MERGE`.
 
 ### Breaking Changes